requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-filter==23.2
django-guardian==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.4
Pillow==10.0.0
Pygments==2.15.1
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1